seperate simple calc into containers and component
