+++ src/App.js
@@ -2,7 +2,7 @@ import './App.css';
 import {
   Switch, Route, Redirect,
 } from 'react-router-dom';
-import SimpleCalc from './components/SimpleCalc';
+import SimpleCalcHome from './containers/SimpleCalcHome';
 import Nav from './components/Nav/Nav';
 import Footer from './components/Footer/Footer';
 import ComplexCalcHome from './containers/ComplexCalcHome';
@@ -13,7 +13,7 @@ function App() {
       <Nav />
 
       <Switch>
-        <Route path="/manual" exact component={SimpleCalc} />
+        <Route path="/manual" exact component={SimpleCalcHome} />
         <Route path="/" exact component={ComplexCalcHome} />
         <Redirect to="/" />
       </Switch>
